requirements.txt
groq>=0.9.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-pptx>=0.6.23
lxml>=5.0.0
rich>=13.0.0
playwright>=1.44.0
python-dotenv>=1.2.1